#36225: remove fix for -0.0 == 0.0 case
