+++ tests/ttnn/unit_tests/operations/eltwise/test_binaryng_fp32.py
@@ -7,6 +7,7 @@ import ttnn
 
 import pytest
 from tests.ttnn.utils_for_testing import assert_with_pcc, assert_with_ulp, assert_allclose
+from models.common.utility_functions import is_blackhole
 
 pytestmark = pytest.mark.use_module_device
 
@@ -432,3 +433,53 @@ def test_subalpha_fp32(alpha, device):
     z_tt_out = ttnn.subalpha(x_tt, y_tt, alpha)
     assert_with_ulp(z_tt_out, z_torch)
     assert_with_pcc(ttnn.to_torch(z_tt_out), z_torch)
+
+
+@pytest.mark.parametrize(
+    "op_name",
+    [
+        "eq",
+    ],
+)
+@pytest.mark.parametrize(
+    "dtype",
+    [
+        "float32",
+    ],
+)
+def test_special_values(device, op_name, dtype):
+    """
+    Comprehensive test for special floating-point values: 0, -0, inf, -inf, nan
+    Tests all combinations of these values as inputs to a binary operation.
+    """
+    torch_fn = getattr(torch, op_name)
+    ttnn_fn = getattr(ttnn, op_name)
+
+    torch_dtype = getattr(torch, dtype)
+    ttnn_dtype = getattr(ttnn, dtype)
+
+    # Special values to test
+    special_values = [0.0, float("inf"), float("-inf"), float("nan"), 1.0, -1.0, -0.0]
+
+    # Create all combinations
+    x_vals = [x for x in special_values for _ in special_values]
+    y_vals = [y for _ in special_values for y in special_values]
+
+    x_torch = torch.tensor(x_vals, dtype=torch_dtype)
+    y_torch = torch.tensor(y_vals, dtype=torch_dtype)
+    z_torch = torch_fn(x_torch, y_torch)
+
+    x_tt = ttnn.from_torch(x_torch, dtype=ttnn_dtype, layout=ttnn.TILE_LAYOUT, device=device)
+    y_tt = ttnn.from_torch(y_torch, dtype=ttnn_dtype, layout=ttnn.TILE_LAYOUT, device=device)
+    z_tt = ttnn_fn(
+        x_tt,
+        y_tt,
+    )
+    tt_out = ttnn.to_torch(z_tt)
+
+    # Note: In Blackhole, (-0.0 == 0.0) returns false
+    if is_blackhole():
+        mask = torch.logical_and(torch.eq(x_torch, -0.0), torch.eq(y_torch, 0.0))
+        tt_out[mask] = 1.0
+
+    assert torch.equal(z_torch, tt_out), "Mismatches found"
